edx_builder/requirements.txt
# EdxBuilder Dependencies

# List of modules
git+https://github.com/Yomguithereal/colifrapy.git
Markdown
lxml
pyyaml